requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
Appium-Python-Client==1.0.2
atomicwrites==1.3.0
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
idna==2.8
importlib-metadata==0.23
iniconfig==1.0.1
more-itertools==7.2.0
MouseInfo==0.1.2
packaging==19.2
Pillow==6.2.1
pluggy==0.13.0
py==1.8.0
PyAutoGUI==0.9.48
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyparsing==2.4.2
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.25
pytest==5.2.1
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-repeat==0.8.0
pytest-rerunfailures==8.0
pytest-sugar==0.9.2
PyTweening==1.0.3
pywin32==228
PyYAML==5.1.2
requests==2.22.0
selenium==3.141.0
six==1.15.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
